+++ backend/packages/caching/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @cat-factory/caching
 
+## 0.4.8
+
+### Patch Changes
+
+- Updated dependencies [2e4d883]
+  - @cat-factory/kernel@0.91.0
+
+## 0.4.7
+
+### Patch Changes
+
+- Updated dependencies [773695b]
+  - @cat-factory/kernel@0.90.0
+
+## 0.4.6
+
+### Patch Changes
+
+- @cat-factory/kernel@0.89.1
+
 ## 0.4.5
 
 ### Patch Changes