+++ backend/packages/eks/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @cat-factory/eks
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [2e4d883]
+  - @cat-factory/contracts@0.101.0
+  - @cat-factory/kernel@0.91.0
+  - @cat-factory/integrations@0.67.1
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [773695b]
+  - @cat-factory/contracts@0.100.0
+  - @cat-factory/kernel@0.90.0
+  - @cat-factory/integrations@0.67.0
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [3981bbb]
+  - @cat-factory/contracts@0.99.0
+  - @cat-factory/integrations@0.66.1
+  - @cat-factory/kernel@0.89.1
+
 ## 0.1.0
 
 ### Minor Changes