+++ backend/packages/provider-cloudflare/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @cat-factory/provider-cloudflare
 
+## 0.7.148
+
+### Patch Changes
+
+- Updated dependencies [2e4d883]
+  - @cat-factory/kernel@0.91.0
+  - @cat-factory/agents@0.38.1
+
+## 0.7.147
+
+### Patch Changes
+
+- Updated dependencies [773695b]
+  - @cat-factory/kernel@0.90.0
+  - @cat-factory/agents@0.38.0
+
+## 0.7.146
+
+### Patch Changes
+
+- @cat-factory/agents@0.37.2
+- @cat-factory/kernel@0.89.1
+
 ## 0.7.145
 
 ### Patch Changes