+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @cat-factory/provider-s3
 
+## 0.2.89
+
+### Patch Changes
+
+- Updated dependencies [2e4d883]
+  - @cat-factory/kernel@0.91.0
+
+## 0.2.88
+
+### Patch Changes
+
+- Updated dependencies [773695b]
+  - @cat-factory/kernel@0.90.0
+
+## 0.2.87
+
+### Patch Changes
+
+- @cat-factory/kernel@0.89.1
+
 ## 0.2.86
 
 ### Patch Changes